Tools/cut_ai/requirements.txt
torch==2.2.0 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.17.0 --index-url https://download.pytorch.org/whl/cpu
opencv-python-headless==4.10.0.84
Pillow==11.0.0
numpy==1.26.4
pika==1.3.2
pytz>=2023.3
